requirements.txt
flask==3.0.0
pymssql==2.2.11
python-dotenv==1.0.0
